refactor: useCreateEventFeedback
