+++ src/modules/createEventFeedback/CreateEventFeedbackPage.vue
@@ -10,6 +10,7 @@
       >
         <CategoryInput
           :question="'Category' + ' ' + (gIndex + 1)"
+          :questionId="group.seq"
           class="mb-1"
           @user-input="handleUserInput(gIndex, $event)"
           @remove="popCategory(gIndex)"
@@ -67,20 +68,13 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, reactive } from "vue";
-import { useRoute } from "vue-router";
-import {
-  AnswerType,
-  QuestionGroupType,
-  SetEventQuestionsInput,
-  SetEventQuestionsQuestionGroupInput,
-  SetEventQuestionsQuestionInput
-} from "@/apollo/types";
+import { defineComponent } from "vue";
 import BaseButton from "@/commons/UI/BaseButton.vue";
 import CategoryInput from "./category-input/CategoryInput.vue";
 import QuestionTextPreview from "./question-text-preview/QuestionTextPreview.vue";
 import QuestionRadioPreview from "./question-radio-preview/QuestionRadioPreview.vue";
 import QuestionStarPreview from "./question-star-preview/QuestionStarPreview.vue";
+import useCreateEventFeedback from "./use-create-event-feedback";
 
 export default defineComponent({
   name: "CreateFeedback",
@@ -92,82 +86,20 @@ export default defineComponent({
     QuestionStarPreview
   },
   setup() {
-    const hello = "hello";
-    let sq = 0;
-    let sqq = 0;
-    const route = useRoute();
-    const eventID = Number(route.params.id);
-    const questionInput = reactive({} as SetEventQuestionsInput);
-    const questionGroups = reactive(
-      [] as SetEventQuestionsQuestionGroupInput[]
-    );
-    const questions = reactive([] as SetEventQuestionsQuestionInput[]);
-    questionInput.eventId = eventID;
-    questionInput.questionGroups = questionGroups;
-
-    const addCategory = () => {
-      sq++;
-      questionGroups.push({
-        type: QuestionGroupType.PostEvent,
-        seq: sq,
-        title: "",
-        questions: [] as SetEventQuestionsQuestionInput[]
-      });
-      console.log(questionGroups);
-    };
-
-    const popCategory = (index: number) => {
-      questionGroups.splice(index, 1);
-      console.log(questionGroups, "remove");
-    };
-
-    const handleUserInput = (index: number, answer: string) => {
-      questionGroups[index].title = answer;
-      console.log(questionGroups);
-    };
-
-    const checkQuestionTypeScale = (type: string) => {
-      return type === AnswerType.Scale;
-    };
-    const addTextQuestion = (index: number) => {
-      sqq++;
-      questionGroups[index].questions.push({
-        seq: sqq,
-        answerType: AnswerType.Text,
-        isOptional: true,
-        title: "",
-        subtitle: ""
-      });
-      console.log(questionGroups);
-    };
-    const addScaleQuestion = (index: number) => {
-      sqq++;
-      questionGroups[index].questions.push({
-        seq: sqq,
-        answerType: AnswerType.Scale,
-        isOptional: true,
-        title: "",
-        subtitle: ""
-      });
-      console.log(questionGroups);
-    };
-
-    const popQuestion = (gIndex: number, index: number) => {
-      questionGroups[gIndex].questions.splice(index, 1);
-      console.log(questions, "delete");
-    };
-
-    const handleQuestionInput = (
-      gIndex: number,
-      index: number,
-      answer: string
-    ) => {
-      questionGroups[gIndex].questions[index].title = answer;
-      console.log(questionGroups);
-    };
+    const {
+      eventID,
+      addCategory,
+      questionGroups,
+      popCategory,
+      handleUserInput,
+      checkQuestionTypeScale,
+      addTextQuestion,
+      addScaleQuestion,
+      popQuestion,
+      handleQuestionInput
+    } = useCreateEventFeedback();
 
     return {
-      hello,
       eventID,
       addCategory,
       questionGroups,
